# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Mycity test API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/client/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticazione client",
                "description": "Effettua il login per un client e restituisce un token",
                "operationId": "8dd833fca3b3f090c3b6ab76b259c039",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "secret_id"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "client_1"
                                    },
                                    "secret_id": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret_1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Auth effettuato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abcdefghijklmnopqrstuvwxyz123456"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-08-02T22:06:31.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenziali non valide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/client/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Effettua il logout del client",
                "operationId": "3fc3698d1cff06f3d744cfc51bf41b09",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Logout effettuato con successo",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/client/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get user profile",
                "description": "Get profile",
                "operationId": "get_profile_client",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/user/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticazione utente",
                "description": "Effettua il login per un utente e restituisce un token",
                "operationId": "489b863a0a302d8ef4671147ad1a87ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@admin.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Auth effettuato con successo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abcdefghijklmnopqrstuvwxyz123456"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-08-02T22:06:31.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenziali non valide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/user/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Effettua il logout dello user",
                "operationId": "7399314579b68143cbb391c26b8c64ba",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Logout effettuato con successo",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/auth/user/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get user profile",
                "description": "Get profile",
                "operationId": "get_user_client",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Get list of personas",
                "description": "Returns list of personas",
                "operationId": "getPersonasList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Persona"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Create new persona",
                "description": "Creates a new persona",
                "operationId": "createPersona",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Persona"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Get persona by ID",
                "description": "Returns a single persona",
                "operationId": "getPersonaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Update existing persona",
                "description": "Updates an existing persona",
                "operationId": "updatePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Persona"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Persona"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Delete persona",
                "description": "Deletes a persona",
                "operationId": "deletePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona deleted successfully"
                    },
                    "404": {
                        "description": "Persona not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Persona": {
                "title": "Persona",
                "required": [
                    "nome",
                    "cognome",
                    "data_di_nascita",
                    "email",
                    "telefono",
                    "codice_fiscale"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "cognome": {
                        "type": "string"
                    },
                    "data_di_nascita": {
                        "type": "string",
                        "format": "date"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "numero": {
                        "type": "string"
                    },
                    "codice_fiscale": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "JWT Bearer Token",
                "name": "Authorization",
                "in": "header"
            },
            "Sanctum": {
                "type": "apiKey",
                "description": "Sanctum Token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth users"
        },
        {
            "name": "Personas",
            "description": "Gestione delle persone"
        }
    ]
}